Fix bugs
- The ledger parser no longer ignores the last transaction in a ledger
journal just because it isn't followed by two newlines.
- Node.js parses dates the US way (mm/dd) while Lloyds incorrectly
exports QIF files the proper way (dd/mm). QIF is badly documented, but
the "correct" way is likely the US way. But since I'm the one who will
mostly be using this, logify now assumes the international way and
converts it to US to work with Node.js.